perf: home rsc optimization, moved videos to client components
