+++ src/app/[locale]/components/MainPage/mainpage.css
@@ -124,112 +124,3 @@
     background-image: url("../../../../../public/assets/images/imgPrevia4_mob.png");
   }
 }
-
-/*Video*/
-.videoPresentacion {
-  display: flex;
-  flex-direction: row;
-  background-color: #276d70;
-  padding: 4em;
-  gap: 2em;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-}
-
-.videoHome {
-  width: 100%;
-  height: auto;
-}
-
-.videoText {
-  color: var(--white);
-}
-
-.videoText h3 {
-  font-size: 22px;
-  font-family: var(--dm-sans);
-  color: #0b273a;
-}
-
-.videoText h1 {
-  font-size: 30px;
-  font-family: var(--montserrat);
-  font-weight: 700;
-}
-
-.videoText p {
-  font-size: 16px;
-  font-family: var(--montserrat);
-}
-
-@media screen and (max-width: 768px) {
-  .videoPresentacion {
-    padding: 0em;
-    gap: 1em;
-  }
-
-  .videoText {
-    padding: 1.5em;
-  }
-}
-
-/*Phrase*/
-
-.phrase {
-  padding: 3em;
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: column;
-  background-image: url("../../../../../public/assets/images/imgTrees.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  color: var(--white);
-}
-
-.phrase h2,
-.phrase p {
-  font-family: var(--montserrat);
-}
-
-.phrase button {
-  max-width: 150px;
-  padding: 1em;
-  border-radius: 40px;
-  border: 0;
-  background-color: var(--strong-green);
-  color: var(--white);
-  font-family: var(--montserrat);
-  font-weight: 500;
-  font-size: 16px;
-}
-
-.phrase button:hover {
-  cursor: pointer;
-  background-color: #276d70;
-}
-
-/*Tamaños Mobile*/
-@media screen and (max-width: 768px) {
-  .layoutHome h1 {
-    font-size: 50px;
-    padding-top: 2em;
-  }
-
-  .layoutHome h3,
-  .servicesText h3 {
-    font-size: 26px;
-  }
-
-  .servicesText h1 {
-    font-size: 36px;
-  }
-
-  .servicesText p {
-    font-size: 15px;
-  }
-
-  .phrase h2 {
-    font-size: 22px;
-  }
-}
