+++ src/app/[locale]/layout.js
@@ -1,18 +1,17 @@
-import './globals.css'
-import RootLayout from './layout-ga';
+import "./globals.css";
+import RootLayout from "./layout-ga";
 
-const { NextIntlClientProvider } = require('next-intl');
-const { notFound } = require('next/navigation');
+const { NextIntlClientProvider } = require("next-intl");
+const { notFound } = require("next/navigation");
 
 export const metadata = {
-  title: 'Oxygen Token',
-  description: 'Landing page Oxygen',
-  icon: '../favicon.ico'
-}
- 
+  title: "Oxygen Token",
+  description: "Landing page Oxygen",
+  icon: "../favicon.ico",
+};
 
 function generateStaticParams() {
-  return [{ locale: 'es' }, { locale: 'en' }];
+  return [{ locale: "es" }, { locale: "en" }];
 }
 
 async function LocaleLayout({ children, params: { locale } }) {
@@ -26,11 +25,10 @@ async function LocaleLayout({ children, params: { locale } }) {
   return (
     <html lang={locale}>
       <body>
-        <RootLayout/>
+        <RootLayout />
         <NextIntlClientProvider locale={locale} messages={messages}>
           {children}
         </NextIntlClientProvider>
-        
       </body>
     </html>
   );
